pyproject.toml
[tool.poetry]
name = "say-bot"
version = "0.1.0"
description = ""
authors = ["Julia <juliariety@gmail.com>"]
license = "0BSD"

[tool.poetry.dependencies]
python = "^3.11"
mastodon-py = "^1.6.1"
beautifulsoup4 = "^4.11.1"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
isort = "^5.10.1"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
